chore: cleanup benchmark
