+++ benchmark/dart_crypto/lib/algorithms.dart
@@ -5,8 +5,9 @@ import 'dart:typed_data';
 import 'package:at_chops/at_chops.dart';
 import 'package:dart_crypto/atchops_better.dart';
 import 'package:dart_crypto/atchops_boring.dart';
-import 'package:dart_crypto/atchops_current.dart';
-import 'package:dart_crypto/atchops_ffi.dart';
+import 'package:dart_crypto/atchops_crypto.dart';
+import 'package:dart_crypto/atchops_encrypt.dart';
+import 'package:dart_crypto/atchops_mbedtls.dart';
 
 abstract class AsyncAtEncryptionAlgorithm {
   FutureOr<Uint8List> encrypt(Uint8List plainData);
@@ -23,21 +24,24 @@ abstract class AsyncSymmetricEncryptionAlgorithm
 }
 
 class Algorithms {
-  final AsyncSymmetricEncryptionAlgorithm atChopsCurrent;
-  final AsyncSymmetricEncryptionAlgorithm atChopsFfi;
-  final AsyncSymmetricEncryptionAlgorithm atChopsBoring;
+  final AsyncSymmetricEncryptionAlgorithm atChopsEncrypt;
   final AsyncSymmetricEncryptionAlgorithm atChopsBetter;
+  final AsyncSymmetricEncryptionAlgorithm atChopsBoring;
+  final AsyncSymmetricEncryptionAlgorithm atChopsCrypto;
+  final AsyncSymmetricEncryptionAlgorithm atChopsMbedtls;
 
   Algorithms(AESKey key)
-      : atChopsCurrent = AtChopsCurrent(key),
-        atChopsFfi = AtChopsFfi(base64.decode(key.key)),
+      : atChopsBetter = AtChopsBetter(base64.decode(key.key)),
         atChopsBoring = AtChopsBoring(base64.decode(key.key)),
-        atChopsBetter = AtChopsBetter(base64.decode(key.key));
+        atChopsCrypto = AtChopsCrypto(base64.decode(key.key)),
+        atChopsEncrypt = AtChopsEncrypt(key),
+        atChopsMbedtls = AtChopsMbedtls(base64.decode(key.key));
 
   Map<String, AsyncSymmetricEncryptionAlgorithm> get asMap => {
-        "current": atChopsCurrent,
         "better": atChopsBetter,
         "boring": atChopsBoring,
-        "mbedtls": atChopsFfi,
+        "crypto": atChopsCrypto,
+        "encrypt": atChopsEncrypt,
+        // "mbedtls": atChopsMbedtls, // ffi stability requires more work, hence disabled
       };
 }
