+++ benchmark/dart_crypto/lib/atchops_mbedtls.dart
@@ -1,129 +1,108 @@
-import 'dart:convert';
+import 'dart:async';
+import 'dart:io';
 
+import 'package:atsdk_ffi/atsdk_ffi.dart';
 import 'package:dart_crypto/algorithms.dart';
-import 'package:dart_crypto/atchops_ffi_old.dart';
 import 'package:ffi/ffi.dart';
 import 'dart:ffi';
 import 'dart:typed_data';
 
+import 'package:path/path.dart' as path;
 import 'package:at_chops/at_chops.dart';
+
+final _libExtension = {
+  'linux': 'so',
+  'macos': 'dylib',
+  'windows': 'dll',
+}[Platform.operatingSystem];
+final _libPath = path.join(Directory.current.path, "../../ffi/dart/build",
+    'libatchops.$_libExtension');
+final _dylib = DynamicLibrary.open(_libPath);
+
+extension on Uint8List {
+  Pointer<T> toPointer<T extends NativeType>() {
+    final Pointer<Uint8> pointer = malloc(length * sizeOf<Uint8>());
+    for (var i = 0; i < length; i++) {
+      (pointer + i).value = this[i];
+    }
+    return Pointer.fromAddress(pointer.address);
+  }
+}
 
 class AtChopsMbedtls extends AsyncSymmetricEncryptionAlgorithm {
   final Uint8List _key;
-  AtChopsMbedtls(AESKey key) : _key = base64.decode(key.key);
-  // final AESKey _key;
-  // AtChopsMbedtls(this._key);
+  final AtSdkFfi ffi = AtSdkFfi(_dylib);
+  AtChopsMbedtls(this._key);
+
   @override
-  Uint8List decrypt(Uint8List encryptedData, {InitialisationVector? iv}) {
-    assert(iv != null);
-
-    final nativeIv =
-        String.fromCharCodes(iv!.ivBytes).toNativeUtf8(); // allocated
-    final key = String.fromCharCodes(_key).toNativeUtf8(); // allocated
-    final keyLength = 256;
-
-    Pointer<Utf8> cipherText =
-        String.fromCharCodes(encryptedData).toNativeUtf8();
-    final cipherTextLength = encryptedData.length;
-    Pointer<UnsignedLong> outLen =
-        malloc.allocate(sizeOf<UnsignedLong>()); // allocated
-
-    int bufSize = cipherTextLength ~/ 3;
-    Pointer<UnsignedChar> buffer =
-        malloc.allocate(bufSize * sizeOf<UnsignedChar>()); // allocated
-
-    int res = AtChopsFfi.decrypt(key, keyLength, nativeIv, cipherText,
-        cipherTextLength, buffer, bufSize, outLen);
-
-    malloc.free(nativeIv);
-    malloc.free(key);
-    malloc.free(cipherText);
-    if (res != 0) {
-      malloc.free(outLen);
-      print("atchops failed encode: $res");
-      return Uint8List(0);
-    }
-
-    bufSize = outLen.value;
-    final decodedTextSize = bufSize;
-    Pointer<Utf8> decodedText =
-        malloc.allocate(decodedTextSize * sizeOf<UnsignedChar>()); // allocated
-    res = AtChopsFfi.decode(
-      buffer,
-      bufSize,
-      decodedText,
-      decodedTextSize,
-      outLen,
-    );
-
-    malloc.free(buffer);
-    if (res != 0) {
-      malloc.free(decodedText);
-      malloc.free(outLen);
-      print("atchops failed decode: $res");
-      return Uint8List(0);
-    }
-
-    String output = decodedText.toDartString(length: outLen.value);
-
-    malloc.free(decodedText);
-    malloc.free(outLen);
-
-    return base64Decode(output);
+  FutureOr<Uint8List> decrypt(Uint8List encryptedData,
+      {InitialisationVector? iv}) {
+    return _crypt(encryptedData, iv: iv!, encrypt: false);
   }
 
   @override
-  Uint8List encrypt(Uint8List plainData, {InitialisationVector? iv}) {
-    assert(iv != null);
-    final encoded = String.fromCharCodes(plainData);
-    final inputText = encoded.toNativeUtf8(); // allocated
-    final inputTextLength = plainData.length;
-    final encodedTextSize = inputTextLength * 4 ~/ 3 + 4;
-    Pointer<Utf8> encodedText =
-        malloc.allocate(encodedTextSize * sizeOf<UnsignedChar>()); // allocated
-    Pointer<UnsignedLong> outLen =
-        malloc.allocate(sizeOf<UnsignedLong>()); // allocated
-    int res = AtChopsFfi.encode(
-      inputText,
-      inputTextLength,
-      encodedText,
-      encodedTextSize,
-      outLen,
-    );
+  FutureOr<Uint8List> encrypt(Uint8List plainData,
+      {InitialisationVector? iv}) async {
+    return await _crypt(Uint8List.fromList(plainData), iv: iv!, encrypt: true);
+  }
 
-    malloc.free(inputText);
-    if (res != 0) {
-      malloc.free(encodedText);
-      malloc.free(outLen);
-      print("atchops failed encode: $res");
-      return Uint8List(0);
+  FutureOr<Uint8List> _crypt(Uint8List plainData,
+      {required InitialisationVector iv, required bool encrypt}) {
+    final nativeIv = iv.ivBytes.toPointer<UnsignedChar>(); // allocated
+    final key = _key.toPointer<UnsignedChar>(); // allocated
+    final keyLength = atchops_aes_size.ATCHOPS_AES_256;
+
+    final bufferLength1 = plainData.length;
+    final buffer1 = plainData.toPointer<UnsignedChar>();
+    late int bufferSize2;
+    if (encrypt) {
+      bufferSize2 = ffi.atchops_aes_ctr_ciphertext_size(bufferLength1);
+    } else {
+      bufferSize2 = ffi.atchops_aes_ctr_plaintext_size(bufferLength1);
     }
-
-    final nativeIv =
-        String.fromCharCodes(iv!.ivBytes).toNativeUtf8(); // allocated
-    final key = String.fromCharCodes(_key).toNativeUtf8(); // allocated
-    final keyLength = 256;
-    int bufSize = outLen.value * 3;
-    Pointer<UnsignedChar> buffer =
-        malloc.allocate(bufSize * sizeOf<UnsignedChar>()); // allocated
-    res = AtChopsFfi.encrypt(key, keyLength, nativeIv, encodedText,
-        inputTextLength, buffer, bufSize, outLen);
+    Pointer<UnsignedChar> buffer2 =
+        malloc.allocate(bufferSize2 * sizeOf<UnsignedChar>()); // allocated
+    Pointer<Size> outLen = malloc.allocate(sizeOf<UnsignedLong>()); // allocated
+    late int res;
+    if (encrypt) {
+      res = ffi.atchops_aes_ctr_encrypt(
+        key,
+        keyLength,
+        nativeIv,
+        buffer1,
+        bufferLength1,
+        buffer2,
+        bufferSize2,
+        outLen,
+      );
+    } else {
+      res = ffi.atchops_aes_ctr_decrypt(
+        key,
+        keyLength,
+        nativeIv,
+        buffer1,
+        bufferLength1,
+        buffer2,
+        bufferSize2,
+        outLen,
+      );
+    }
     malloc.free(nativeIv);
     malloc.free(key);
-    malloc.free(encodedText);
-
+    malloc.free(buffer1);
     if (res != 0) {
-      malloc.free(buffer);
+      malloc.free(buffer2);
       malloc.free(outLen);
-      print("atchops failed crypt: $res");
+
+      // String prefix = (encrypt) ? "en" : "de";
+      // print("atchops failed ${prefix}crypt: $res");
       return Uint8List(0);
     }
-
-    Pointer<Uint8> p = Pointer.fromAddress(buffer.address);
-    final output = Uint8List.fromList(p.asTypedList(outLen.value));
-
-    malloc.free(buffer);
+    final output =
+        Pointer<Uint8>.fromAddress(buffer2.address).asTypedList(outLen.value);
+    malloc.free(buffer2);
     malloc.free(outLen);
+
     return output;
   }
 }
